# IdleSchemes.Api/openapispec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "IdleSchemes.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/registrations/start": {
      "post": {
        "tags": [
          "Registration"
        ],
        "operationId": "StartRegistration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRegistrationModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRegistrationModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StartRegistrationModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationInfoModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationInfoModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationInfoModel"
                }
              }
            }
          }
        }
      }
    },
    "/registrations/reserve": {
      "post": {
        "tags": [
          "Registration"
        ],
        "operationId": "ReserveTickets",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveTicketsModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveTicketsModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveTicketsModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationInfoModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationInfoModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationInfoModel"
                }
              }
            }
          }
        }
      }
    },
    "/registrations/confirm": {
      "post": {
        "tags": [
          "Registration"
        ],
        "operationId": "ConfirmPayment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRegistrationModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRegistrationModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRegistrationModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmRegistrationModel": {
        "type": "object",
        "properties": {
          "registrationId": {
            "type": "string"
          },
          "registrationSecret": {
            "type": "string"
          }
        }
      },
...
      "RegistrationInfoModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          },
          "started": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ReservationInfoModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          },
          "tickets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ticket"
            }
          }
        }
      },
      "ReserveTicketsModel": {
        "type": "object",
        "properties": {
          "registrationId": {
            "type": "string"
          },
          "registrationSecret": {
            "type": "string"
          },
          "ticketReservations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TicketReservation"
            }
          }
        }
      },
      "SeatAssignment": {
        "type": "object",
        "properties": {
          "seat": {
            "type": "string",
            "nullable": true
          },
          "holderName": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "StartRegistrationModel": {
        "type": "object",
        "properties": {
          "instanceId": {
            "type": "string"
          }
        }
      },
      "Ticket": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "class": {
            "type": "string"
          },
          "seat": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TicketReservation": {
        "type": "object",
        "properties": {
          "class": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "count": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "seats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeatAssignment"
            },
            "nullable": true
          },
          "holderName": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  },
...
}